Fix layer rotating black lines inbetween tiles
git-svn-id: file:///desktop/SourceForge%20Backup/sphere/svnrepos/trunk@1126 c0eb2ce6-7f40-480a-a1e0-6c4aea08c2c2

+++ sphere/source/engine/rendersort.cpp
@@ -25,7 +25,7 @@ CRenderSort::AddObject(int draw_x, int draw_y, int sort_y, int draw_w, int draw_
 
 ////////////////////////////////////////////////////////////////////////////////
 
-void CalculateRotateBlitPoints(int tx[4], int ty[4], int x, int y, int w, int h, double radians);
+void CalculateRotateBlitPoints(int tx[4], int ty[4], double x, double y, double w, double h, double radians);
 
 void
 CRenderSort::DrawObjects()
